refactor(svelte): split surface-inspection pure unit tests by module
Replace the 941-line surface-inspection.test.ts monolith with co-located
suites for frame, apply, and teardown decision tables. Assertions preserved.
